🧩 Separated ForecastListView from ForecastBlocBuilder widget ✅ Resolved TODO note in ForecastBlocBuilder 🎨 Changed map color
